Dj_TPS/pyproject.toml
[tool.poetry]
name = "TPS"
version = "0.1.0"
description = "Take Photo ERP System"
authors = ["Александр Кривов"]
packages = [{include = "Dj_TPS"}, {include = "tph_system"}, {include = "users"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-filter = "^24.3"
django-axes = "^7.0.1"
django-debug-toolbar = "^4.4.6"
django-background-tasks = "^1.2.8"
psycopg2 = "^2.9.10"
icalendar = "^6.1.0"
python-dotenv = "^1.0.1"
qrcode = "^8.0"
pillow = "^11.1.0"
djangorestframework = "^3.15.2"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
aiogram = "^3.18.0"
httpx = "^0.28.1"
djangorestframework-simplejwt = "^5.5.0"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
